Don't include pthread.h and sched.h directly
Add pte_pthread.h and pte_shed.h to wrap the respective PTE headers. The
idea is that in the future these two files could contain some macros to
decide which includes to load, or setup some defines to alter the
behaviour of the included pthread.h and sched.h files.  This is the
first step to allow us to use the toolchain-provided headers.

We move the inclusion of sched.h and pthread.h from the source files to
implement.h, since this file needs those definitions anyway and is
included by most source files.